+++ src/index.js
@@ -15,5 +15,6 @@ export * from './reducer'
 export * from './selectors'
 export * from './actions'
 export * from './hooks/index'
+export * from './mutations/helpers'
 export { default as combineRjs } from './combineRjs'
 export { default as rjMiddleware } from './rjMiddleware'
